Gets date from Firebase
