--- ArtSpacePrototype/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  ArtSpacePrototype
-//
-//  Created by God on 1/21/20.
-//  Copyright © 2020 God. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
