+++ ArtSpacePrototype/ViewControllers/Views/DetailVC.swift
